tantivy-py: Upgrade PyO3.
This removes our nightly requirement but sadly it adds a requirement for
things that are kept inside a Python class to be Send. Luckily for us
almost everything in Tantivy is Send, except for the Query trait.

This patch works around this by keeping the parser and query string
inside our python Query object. This sadly means that we are going to
parse the query string twice.